Change reputation service to use a predefined 5-star system

+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/message_container/list/ChatMessageListItem.java
@@ -322,7 +322,7 @@ public final class ChatMessageListItem<M extends ChatMessage, C extends ChatChan
         return isBisqEasyPublicChatMessageWithPeerOffer() && hasBisqEasyOfferWithDirection(Direction.SELL);
     }
 
-    public int getReputationStarCount() {
+    public double getReputationStarCount() {
         return reputationScoreDisplay.getNumberOfStars();
     }
